requirements.txt
Django==2.0
django-cors-headers==2.2.0
djangorestframework==3.7.7
gunicorn==19.7.1
pytz==2018.3
whitenoise==3.3.1
requests
psycopg2
